Make sure FVPropValuPerSubdomainMaterial works outside of FV contexts
E.g. when adding the postprocessor for flow rates I was triggering
assertions in this material class that the FaceInfo was null because it
was being called in the user object thread before I could call
`setFaceInfo`. So now we actively get the face info when we have
non-null current element.