Add y-channel friction + porosity jumps using continuous function
Here's my newest technique for handling porosity jumps: make the jump
into a continuous function in which the change happens over the two
elements neighboring the "jump" face. Then we can use a function to
exactly capture this gradient in the PNSFVPGradEpsilon
FVElementalKernel. Tackling the porosity "jump" this way allows me to
correctly capture pressure/density behavior when applying the
epsilon 1-0.5-0.25-0.5-1 staged jumps test: the final pressure ends up
being equal to the inlet pressure. I do not see this if I try to treat
the porosity jump discontinuously with an auxiliary variable. In that
case I see a pressure that gets lower with every porosity change, *even*
when increasing the porosity which isn't right